consolidate watch paths logic
