+++ packages/app/src/cli/models/extensions/extension-instance.test.ts
@@ -1,5 +1,6 @@
 import {testFunctionExtension} from '../app/app.test-data.js'
 import {FunctionConfigType} from '../extensions/specifications/function.js'
+import {joinPath} from '@shopify/cli-kit/node/path'
 import {describe, expect, test} from 'vitest'
 
 function functionConfiguration(): FunctionConfigType {
@@ -17,28 +18,64 @@ describe('watchPaths', async () => {
   test('returns an array for a single path', async () => {
     const config = functionConfiguration()
     config.build = {
-      watch: 'single-path',
+      watch: 'src/single-path.foo',
     }
     const extensionInstance = await testFunctionExtension({
       config,
+      dir: 'foo',
     })
 
     const got = extensionInstance.watchPaths
 
-    expect(got).toEqual(['single-path'])
+    expect(got).toEqual([joinPath('foo', 'src', 'single-path.foo'), joinPath('foo', '**', 'input*.graphql')])
   })
 
-  test('returns an array for an array', async () => {
+  test('returns default paths for javascript', async () => {
+    const config = functionConfiguration()
+    config.build = {}
+    const extensionInstance = await testFunctionExtension({
+      config,
+      entryPath: 'src/index.js',
+      dir: 'foo',
+    })
+
+    const got = extensionInstance.watchPaths
+
+    expect(got).toEqual([
+      joinPath('foo', 'src', '**', '*.js'),
+      joinPath('foo', 'src', '**', '*.ts'),
+      joinPath('foo', '**', 'input*.graphql'),
+    ])
+  })
+
+  test('returns configured paths and input query', async () => {
     const config = functionConfiguration()
     config.build = {
-      watch: ['one-path', 'two-path', 'red-path', 'blue-path'],
+      watch: ['src/**/*.rs', 'src/**/*.foo'],
     }
+    const extensionInstance = await testFunctionExtension({
+      config,
+      dir: 'foo',
+    })
+
+    const got = extensionInstance.watchPaths
+
+    expect(got).toEqual([
+      joinPath('foo', 'src/**/*.rs'),
+      joinPath('foo', 'src/**/*.foo'),
+      joinPath('foo', '**', 'input*.graphql'),
+    ])
+  })
+
+  test('returns null if not javascript and not configured', async () => {
+    const config = functionConfiguration()
+    config.build = {}
     const extensionInstance = await testFunctionExtension({
       config,
     })
 
     const got = extensionInstance.watchPaths
 
-    expect(got).toEqual(['one-path', 'two-path', 'red-path', 'blue-path'])
+    expect(got).toBeNull()
   })
 })
